Folsom 2012.2.4

+++ cinder/volume/netapp.py
@@ -431,10 +431,17 @@ class NetAppISCSIDriver(driver.ISCSIDriver):
     def _remove_destroy(self, name, project):
         """Remove the LUN from the dataset, also destroying it.
 
-        Remove the LUN from the dataset and destroy the actual LUN on the
-        storage system.
+        Remove the LUN from the dataset and destroy the actual LUN and Qtree
+        on the storage system.
         """
-        lun = self._lookup_lun_for_volume(name, project)
+        try:
+            lun = self._lookup_lun_for_volume(name, project)
+            lun_details = self._get_lun_details(lun.id)
+        except exception.VolumeBackendAPIException:
+            msg = _("No entry in LUN table for volume %s.")
+            LOG.info(msg % name)
+            return
+
         member = self.client.factory.create('DatasetMemberParameter')
         member.ObjectNameOrId = lun.id
         members = self.client.factory.create('ArrayOfDatasetMemberParameter')
@@ -445,11 +452,28 @@ class NetAppISCSIDriver(driver.ISCSIDriver):
         try:
             server.DatasetRemoveMember(EditLockId=lock_id, Destroy=True,
                                        DatasetMemberParameters=members)
+            res = server.DatasetEditCommit(EditLockId=lock_id,
+                                           AssumeConfirmation=True)
+        except (suds.WebFault, Exception):
+            server.DatasetEditRollback(EditLockId=lock_id)
+            msg = _('Failed to remove and delete dataset LUN member')
+            raise exception.VolumeBackendAPIException(data=msg)
+
+        for info in res.JobIds.JobInfo:
+            self._wait_for_job(info.JobId)
+
+        #Note(rushiagr): not possible to delete Qtree & its LUN in one
+        #               transaction
+        member.ObjectNameOrId = lun_details.QtreeId
+        lock_id = server.DatasetEditBegin(DatasetNameOrId=lun.dataset.id)
+        try:
+            server.DatasetRemoveMember(EditLockId=lock_id, Destroy=True,
+                                       DatasetMemberParameters=members)
             server.DatasetEditCommit(EditLockId=lock_id,
                                      AssumeConfirmation=True)
         except (suds.WebFault, Exception):
             server.DatasetEditRollback(EditLockId=lock_id)
-            msg = _('Failed to remove and delete dataset member')
+            msg = _('Failed to remove and delete dataset Qtree member')
             raise exception.VolumeBackendAPIException(data=msg)
 
     def create_volume(self, volume):
@@ -1105,6 +1129,10 @@ class NetAppCmodeISCSIDriver(driver.ISCSIDriver):
         """Driver entry point for destroying existing volumes."""
         name = volume['name']
         handle = self._get_lun_handle(name)
+        if not handle:
+            msg = _("No entry in LUN table for volume %(name)s.")
+            LOG.warn(msg % locals())
+            return
         self.client.service.DestroyLun(Handle=handle)
         LOG.debug(_("Destroyed LUN %s") % handle)
         self.lun_table.pop(name)
@@ -1212,7 +1240,12 @@ class NetAppCmodeISCSIDriver(driver.ISCSIDriver):
 
     def delete_snapshot(self, snapshot):
         """Driver entry point for deleting a snapshot."""
-        handle = self._get_lun_handle(snapshot['name'])
+        name = snapshot['name']
+        handle = self._get_lun_handle(name)
+        if not handle:
+            msg = _("No entry in LUN table for snapshot %s.")
+            LOG.warn(msg % name)
+            return
         self.client.service.DestroyLun(Handle=handle)
         LOG.debug(_("Destroyed LUN %s") % handle)
